+++ virtual-desktop/src/app/i18n/i18n.module.ts
@@ -1,0 +1,41 @@
+/*
+  This program and the accompanying materials are
+  made available under the terms of the Eclipse Public License v2.0 which accompanies
+  this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
+
+  SPDX-License-Identifier: EPL-2.0
+
+  Copyright Contributors to the Zowe Project.
+*/
+
+import { NgModule, APP_INITIALIZER, LOCALE_ID } from '@angular/core';
+import { LanguageLocaleService } from '../shared/language-locale.service';
+import { localeInitializer, localeIdFactory } from './locale-initializer.provider';
+
+
+@NgModule({
+  providers: [
+    { provide: LOCALE_ID, useFactory: localeIdFactory, deps: [LanguageLocaleService]},
+    {
+      provide: APP_INITIALIZER,
+      multi: true,
+      useFactory: localeInitializer,
+      deps: [LanguageLocaleService, LOCALE_ID]
+    }
+  ]
+})
+export class I18nModule {
+
+}
+
+
+/*
+  This program and the accompanying materials are
+  made available under the terms of the Eclipse Public License v2.0 which accompanies
+  this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
+
+  SPDX-License-Identifier: EPL-2.0
+
+  Copyright Contributors to the Zowe Project.
+*/
+
